Splitting up controllers to be more logical. Fixing routes and methods.
