+++ src/AppBundle/Controller/DefaultController.php
@@ -86,312 +86,6 @@ class DefaultController extends Controller
             'sorting' => $sorting
         ]);
     }
-
-    /**
-     * @Route("/comments/new", name="new_comment")
-     */
-    public function newCommentAction(Request $request) 
-    {
-        // Only make the request submission on a POST request
-        if($request->isMethod('POST')) {
-
-            $post_id = $request->get('post_id');
-
-            // Get the Post Number
-            $post = $this->getDoctrine()
-                     ->getRepository('AppBundle:Post')
-                     ->find($post_id);
-
-            // Need to get the current user based on security acces
-            $user = self::getCurrentUser($this);
-
-            // Get the User's IP address
-            $comment_ip = self::getCurrentIp($this);
-        
-            // Get the body of the comment from the request
-            $body = $request->get('body');
-        
-            // We have everything we need now
-            // Time to add the post to the database
-            try {
-                $em = self::getEntityManager();
-                $comment = new Comment;
-                $comment->setPost($post);
-                $comment->setBody($body);
-                $comment->setIpAddress($comment_ip);
-                $comment->setUser($user);
-                $em->persist($comment);
-                $em->flush();
-                return new JsonResponse(array('status' => 200, 'message' => 'Success in posting comments', 'comment_id' => $comment->getId()));
-            } catch (\Doctrine\DBAL\DBALException $e) {
-                return new JsonResponse(array('status' => 400, 'message' => 'Unable to comment.'));
-            }   
-        } else {
-            return $this->render('post/post.html.twig');
-        }
-    }
-    
-    /**
-     * @Route("/posts/new", name="new_post")
-     */
-    public function newPostAction(Request $request) 
-    {
-        // Only make the request submission on a POST request
-        if($request->isMethod('POST')) {
-            // Get the User's IP address
-            $post_ip = self::getCurrentIp($this);
-        
-            // Need to get the current user based on security acces
-            $user = self::getCurrentUser($this);
-        
-            // Get the body of the post from the request
-            $body = $request->get('body');
-        
-            // We have everything we need now
-            // Time to add the post to the database
-            try {
-                $em = self::getEntityManager();
-                $post = new Post;
-                $post->setBody($body);
-                $post->setIpAddress($post_ip);
-                $post->setCollege($user->getCollege());
-                $post->setUser($user);
-                $em->persist($post);
-                $em->flush();
-                return new JsonResponse(array('status' => 200, 'message' => 'Success'));
-            } catch (\Doctrine\DBAL\DBALException $e) {
-                return new JsonResponse(array('status' => 400, 'message' => 'Unable to submit post.'));
-            }   
-        } else {
-            return $this->render('default/new_post.html.twig');
-        }
-    }
-
-    /**
-     * @Route("/posts/{post_id}", name="post_view", requirements={"post_id" = "\d+"})
-     */
-    public function viewPostAction(Request $request, $post_id) 
-    {
-        $user = self::getCurrentUser($this);
-        
-        $em = self::getEntityManager();
-        
-        // Get the post from the post_id in the database
-        $post = $em->getRepository('AppBundle:Post')
-                   ->find($post_id);
-        
-        $like = $em->getRepository('AppBundle:PostLikes')
-                   ->findOneBy(array('post' => $post_id, 'user' => $user->getId()));
-        
-        /*
-        EQUIVALENT QUERY TO BUILDER BELOW
-
-       "SELECT c.id, c.post_id, c.upvotes, 
-                c.downvotes, c.body, c.reports, 
-                p.created, l.is_like, l.user_id, 
-                l.comment_id
-         FROM comments c
-         WHERE c.post_id = :postid
-         LEFT JOIN comment_likes l
-         ON c.id = l.comment_id AND l.user_id = ? 
-         GROUP BY c.id, c.post_id, c.body, c.upvotes
-                  c.downvotes, c.reports,
-                  c.created, l.is_like, l.user_id,
-                  l.comment_id
-         ORDER BY created DESC;";
-        
-        */
-        
-        $builder = $em->createQueryBuilder();
-        $builder
-            ->select('c', 'l')
-            ->from('AppBundle:Comment', 'c') 
-            ->where('c.post = :postid AND c.hidden = false')
-            ->setParameter('postid', $post->getId())
-            ->leftJoin(
-                'c.likes',
-                'l',
-                \Doctrine\ORM\Query\Expr\Join::WITH,
-                'c.id = l.comment AND l.user = :user'
-                )
-            ->setParameter('user', $user->getId())
-            ->groupBy('c', 'l')
-            ->orderBy('c.created', 'DESC');
-                
-        $comments = $builder->getQuery()->getResult();
-    
-        // If anything other than a post is returned (including null)
-        // throw an error.
-        if (!$post) {
-            throw $this->createNotFoundException(
-                "Post not found!"
-            );
-        }
-        
-        return $this->render('default/post.html.twig', [
-            'post' => $post, 
-            'like' => $like,
-            'comments' => $comments
-        ]);
-    }
-    
-     /**
-     * @Route("/upvote", name="upvote")
-     * @Method({"POST"})
-     */
-    public function upvoteAction(Request $request) 
-	{
-        // Get post id from the request
-        $post_id = $request->get("post_id");
-        
-        // Get current user
-        $user = self::getCurrentUser($this);
-        
-        // Get the entity manager for Doctrine
-        $em = self::getEntityManager();
-
-		// Get the post from the post_id in the database
-        $post = $em->getRepository('AppBundle:Post')
-                   ->find($post_id);
-    
-        // If anything other than a post is returned (including null)
-        // throw an error.
-        if (!$post) {
-            throw $this->createNotFoundException(
-                'No post found for id ' . $id
-            );
-        }
-
-        try{
-            $like = $em->getRepository('AppBundle:PostLikes')
-                       ->findOneBy(array('post' => $post_id, 'user' => $user->getId()));
-            
-            if(!isset($like)) {
-                $like = new PostLikes;
-                $like->setIsLike(true);
-                $like->setUser(self::getCurrentUser($this));
-                $like->setPost($post);
-                $post->setUpvotes($post->getUpvotes() + 1);
-                $post->addLike($like);
-                $em->persist($like);
-            } else {
-                if($like->getIsLike()) {
-                    $post->setUpvotes($post->getUpvotes() - 1);
-                    $post->removeLike($like);
-                    $em->remove($like);
-                } else {
-                    $post->setUpvotes($post->getUpvotes() + 1);
-                    $post->setDownvotes($post->getDownvotes() - 1);
-                    $like->setIsLike(true);
-                    $em->persist($like);
-                }
-            }
-            $post->setScore(self::hot($post->getUpvotes(), $post->getDownvotes(), $post->getCreated()));
-            $em->persist($post);
-            $em->flush();
-            $score = ($post->getUpvotes() - $post->getDownvotes());
-        } catch (\Docrine\DBAL\DBALException $e) {
-            return new JsonResponse(array('status' => 400, 'message' => 'Unable to add like. $e->message'));
-        }
-        
-        return new JsonResponse(array('status' => 200, 'message' => 'Success on upvote.', 'score' => $score));
-	}
-    
-    /**
-     * @Route("/downvote", name="downvote")
-     * @Method({"POST"})
-     */
-    public function downvoteAction(Request $request) 
-    {
-        // Get the post_id
-        $post_id = $request->get('post_id');
-        
-        // Get current user
-        $user = self::getCurrentUser($this);
-        
-        // Get the entity manager
-        $em = self::getEntityManager();
-        
-        // Get the post from the post_id in the database
-        $post = $em->getRepository('AppBundle:Post')
-                   ->find($post_id);
-        
-        // If anything other than a post is returned (including null)
-        // throw an error.
-        if (!$post) {
-            throw $this->createNotFoundException(
-                'No post found for id ' . $id
-            );
-        }
-
-        // Try to add the downvote
-        try {
-            $like = $em->getRepository('AppBundle:PostLikes')
-                       ->findOneBy(array('post' => $post_id, 'user' => $user->getId()));
-
-            if(!isset($like)) {
-                $dislike = new PostLikes;
-                $dislike->setIsLike(false);
-                $dislike->setUser(self::getCurrentUser($this));
-                $dislike->setPost($post);
-                $post->setDownvotes($post->getDownvotes() + 1);
-                $post->addLike($dislike);
-                $em->persist($dislike);
-            } else {
-                if($like->getIsLike()) {
-                    $post->setUpvotes($post->getUpvotes() - 1);
-                    $post->setDownvotes($post->getDownvotes() + 1);
-                    $like->setIsLike(false);
-                    $em->persist($like);
-                } else {
-                    $post->setDownvotes($post->getDownvotes() - 1);
-                    $em->remove($like);
-                    $post->removeLike($like);
-                }
-            }
-            $post->setScore(self::hot($post->getUpvotes(), $post->getDownvotes(), $post->getCreated()));
-            $em->persist($post);
-            $em->flush();
-            $score = ($post->getUpvotes() - $post->getDownvotes());
-        } catch (\Doctrine\DBAL\DBALException $e) {
-            return new JsonResponse(array('status' => 400, 'message' => 'Unable to dislike. $e->message'));  
-        }
-
-        return new JsonResponse(array('status' => 200, 'message' => 'Success on upvoting', 'score' => $score));
-    }
-    
-    /**
-     * @Route("/remove", name="remove")
-     * @Method({"POST"})
-     */
-    public function removePost(Request $request) 
-    {
-         // Get post id from the request
-        $post_id = $request->get("post_id");
-
-        // Get the post from the post_id in the database
-        $post = $this->getDoctrine()
-                     ->getRepository('AppBundle:Post')
-                     ->find($post_id);
-    
-        // If anything other than a post is returned (including null)
-        // throw an error.
-        if (!$post) {
-            throw $this->createNotFoundException(
-                'No post found for id ' . $id
-            );
-        }
-        
-        // Time to delete the post to the database
-        try {
-            $em = self::getEntityManager();
-            $em->remove($post);
-            $em->flush();
-            return new JsonResponse(array('status' => 200, 'message' => 'Success'));
-        } catch (\Doctrine\DBAL\DBALException $e) {
-            return new JsonResponse(array('status' => 400, 'message' => 'Unable to delete post.'));
-        }   
-    } 
     
     /**
      * @Route("/login", name="login")
@@ -417,79 +111,10 @@ class DefaultController extends Controller
     }
     
     /**
-     * @Route("/confirm/{token}", name="confirm_email")
-     */
-    public function confirmEmailAction(Request $request, $token) {
-        
-        $em = self::getEntityManager();
-        
-        $auth = $em->getRepository("AppBundle:EmailAuth")
-                   ->findOneBy(array('token' => $token, 'verified' => false));
-        
-        if(!$auth) {
-            return $this->render(
-                'security/confirm.html.twig',
-                array(
-                    'error' => 'Token is invalid or has already been used',
-                )
-            );
-        }
-            
-        $user = $em->getRepository("AppBundle:User")
-                   ->find($auth->getUser());
-        
-        if(!$user) {
-            return $this->render(
-                'security/confirm.html.twig',
-                array(
-                    'error' => 'User could not be found. Please contact support at adrestiaweb@gmail.com.',
-                )
-            );
-        }
-        
-        $user->setEmailConfirmed(true);
-        $auth->setVerified(true);
-        $em->persist($user);
-        $em->persist($auth);
-        $em->flush();
-        
-        return $this->render(
-            'security/confirm.html.twig'
-        );
-    }
-    
-    /**
      * @Route("/login_check", name="login_check")
      */
     public function loginCheckAction(Request $request) {
         
-    }
-    
-    /**
-     * @Route("/suffix", name="email_suffix")
-     * @Method({"POST"})
-     */
-    public function suffixAction(Request $request) {
-        $name = $request->get('college');
-        
-        try {
-            $em = self::getEntityManager();
-        
-            $college = $em->getRepository('AppBundle:College')
-                          ->findOneBy(array('name' => $name));
-            
-            if(!$college) {
-                throw $this->createNotFoundException(
-                    'No college found with name ' . $name
-                );
-            }
-            
-            $suffix = $college->getSuffix();
-            
-            return new JsonResponse(array('status' => 200, 'suffix' => $suffix));
-        } catch (\Doctrine\DBAL\DBALException $e) {
-            return new JsonResponse(array('status' => 400, 'message' => 'Unable to get suffix.'));
-        }   
     }
     
     /**
